+++ src/commonTest/kotlin/uk/gov/hmrc/calculator/services/mocks/MockDateTimeService.kt
@@ -18,6 +18,6 @@ package uk.gov.hmrc.calculator.services.mocks
 import com.soywiz.klock.DateTime
 import uk.gov.hmrc.calculator.services.DateTimeService
 
-class MockDateTimeService(val dateTime: DateTime) : DateTimeService {
+class MockDateTimeService(private val dateTime: DateTime) : DateTimeService {
     override fun now(): DateTime = dateTime
 }