+++ src/jvmTest/kotlin/uk/gov/hmrc/calculator/ParameterizedCalculatorTests.kt
@@ -25,6 +25,7 @@ import org.junit.jupiter.params.provider.CsvFileSource
 import uk.gov.hmrc.calculator.model.Country
 import uk.gov.hmrc.calculator.model.PayPeriod
 import uk.gov.hmrc.calculator.model.TaxYear
+import uk.gov.hmrc.calculator.utils.prettyPrintDataClass
 
 internal class ParameterizedCalculatorTests {
 
@@ -34,7 +35,7 @@ internal class ParameterizedCalculatorTests {
         inputTaxCode: String,
         inputWages: Double,
         @ConvertWith(PayPeriodConverter::class) inputPayPeriod: PayPeriod,
-        inputTaxYear: Int,
+        @ConvertWith(TaxYearConverter::class) inputTaxYear: TaxYear,
         inputIsPensionAge: Boolean,
         @ConvertWith(CountryConverter::class) expectedCountry: Country,
         expectedYearlyNiEmployee: Double,
@@ -51,7 +52,7 @@ internal class ParameterizedCalculatorTests {
             taxCode = inputTaxCode,
             wages = inputWages,
             payPeriod = inputPayPeriod,
-            taxYear = TaxYear.fromInt(inputTaxYear),
+            taxYear = inputTaxYear,
             isPensionAge = inputIsPensionAge
         ).run()
 
@@ -84,7 +85,7 @@ internal class ParameterizedCalculatorTests {
         inputTaxCode: String,
         inputWages: Double,
         @ConvertWith(PayPeriodConverter::class) inputPayPeriod: PayPeriod,
-        inputTaxYear: Int,
+        @ConvertWith(TaxYearConverter::class) inputTaxYear: TaxYear,
         inputIsPensionAge: Boolean,
         @ConvertWith(CountryConverter::class) expectedCountry: Country,
         expectedYearlyNiEmployee: Double,
@@ -101,11 +102,9 @@ internal class ParameterizedCalculatorTests {
             taxCode = inputTaxCode,
             wages = inputWages,
             payPeriod = inputPayPeriod,
-            taxYear = TaxYear.fromInt(inputTaxYear),
+            taxYear = inputTaxYear,
             isPensionAge = inputIsPensionAge
         ).run()
-
-        println(response)
 
         assertEquals(expectedCountry, response.country, "Country did not match")
         assertEquals(expectedIsKCode, response.isKCode)
@@ -130,11 +129,65 @@ internal class ParameterizedCalculatorTests {
         assertEquals(PayPeriod.WEEKLY, response.weekly.payPeriod)
     }
 
+    @ParameterizedTest(name = "tax_code={0}, wages={1}, pay_period={2}, tax_year={3}, is_pension_age={4}")
+    @CsvFileSource(resources = ["/data2022_Nov_Revised.csv"], numLinesToSkip = 1)
+    fun `Tax calculations 2022 Nov revised`(
+        inputTaxCode: String,
+        inputWages: Double,
+        @ConvertWith(PayPeriodConverter::class) inputPayPeriod: PayPeriod,
+        @ConvertWith(TaxYearConverter::class) inputTaxYear: TaxYear,
+        inputIsPensionAge: Boolean,
+        @ConvertWith(CountryConverter::class) expectedCountry: Country,
+        expectedYearlyNiEmployee: Double,
+        expectedYearlyNiEmployer: Double,
+        expectedYearlyIncomeTax: Double,
+        expectedYearlyTotalDeduction: Double,
+        expectedYearlyTakeHome: Double,
+        expectedYearlyWages: Double,
+        expectedYearlyTaxFreeAmount: Double,
+        expectedYearlyKCodeAdjustment: Double?,
+        expectedIsKCode: Boolean
+    ) {
+        val response = Calculator(
+            taxCode = inputTaxCode,
+            wages = inputWages,
+            payPeriod = inputPayPeriod,
+            taxYear = inputTaxYear,
+            isPensionAge = inputIsPensionAge
+        ).run()
+
+        assertEquals(expectedCountry, response.country, "Country did not match")
+        assertEquals(expectedIsKCode, response.isKCode)
+
+        val yearlyPeriod = response.yearly
+        println(inputTaxCode)
+        println(yearlyPeriod.prettyPrintDataClass())
+        println("expectedYearlyNiEmployer=$expectedYearlyNiEmployer, yearlyPeriod.employersNI=$yearlyPeriod.employersNI")
+        assertEquals(PayPeriod.YEARLY, yearlyPeriod.payPeriod)
+        assertEquals(expectedYearlyNiEmployee, yearlyPeriod.employeesNI, "Yearly employee NI did not match")
+        assertEquals(expectedYearlyNiEmployer, yearlyPeriod.employersNI, "Yearly employer NI did not match")
+        assertEquals(expectedYearlyIncomeTax, yearlyPeriod.taxToPay, "Yearly income tax did not match")
+        assertEquals(expectedYearlyTotalDeduction, yearlyPeriod.totalDeductions, "Yearly total deductions did not match")
+        assertEquals(expectedYearlyTakeHome, yearlyPeriod.takeHome, "Yearly take home did not match")
+        assertEquals(expectedYearlyWages, yearlyPeriod.wages, "Yearly wages did not match")
+        assertEquals(expectedYearlyTaxFreeAmount, yearlyPeriod.taxFree, "Yearly tax free amount did not match")
+        assertEquals(expectedYearlyKCodeAdjustment, yearlyPeriod.kCodeAdjustment, "Yearly K code adjustment did not match")
+        assertFalse(yearlyPeriod.maxTaxAmountExceeded)
+
+        assertEquals(PayPeriod.MONTHLY, response.monthly.payPeriod)
+        assertEquals(PayPeriod.FOUR_WEEKLY, response.fourWeekly.payPeriod)
+        assertEquals(PayPeriod.WEEKLY, response.weekly.payPeriod)
+    }
+
     class PayPeriodConverter : ArgumentConverter {
         override fun convert(source: Any, context: ParameterContext?) = PayPeriod.valueOf(source as String)
     }
 
     class CountryConverter : ArgumentConverter {
         override fun convert(source: Any, context: ParameterContext?) = Country.valueOf(source as String)
+    }
+
+    class TaxYearConverter : ArgumentConverter {
+        override fun convert(source: Any, context: ParameterContext?) = TaxYear.valueOf(source as String)
     }
 }